Fix cast warning

See merge request BC/public/belle-sip!201

+++ src/transports/udp_listeningpoint.c
@@ -44,7 +44,7 @@ static belle_sip_channel_t *udp_create_channel(belle_sip_listening_point_t *lp, 
 		sock=(int)udp_lp->sock;
 	belle_sip_channel_t *chan=belle_sip_channel_new_udp(lp->stack
 														,sock
-														,belle_sip_uri_get_host(lp->listening_uri)
+														,(int)belle_sip_uri_get_host(lp->listening_uri)
 														,belle_sip_uri_get_port(lp->listening_uri)
 														,hop->host
 														,hop->port, hop->port_is_explicit);
